requirements.txt
# Core Dependencies
streamlit==1.31.0
pandas==2.2.0
numpy<2.0.0

# API & HTTP
requests==2.31.0
stravalib==1.6.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Visualization
plotly==5.18.0
folium==0.15.1
matplotlib==3.8.2
seaborn==0.13.1

# Machine Learning
scikit-learn==1.4.0
prophet==1.1.5

# Utilities
python-dotenv==1.0.1
pytz==2024.1

# Streamlit (additional packages)
streamlit-folium==0.11.0
statsmodels==0.14.0